test(mitm-addon): cover jsonl writer lifecycle behavior
